# application/docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  v1.AccountIdBody:
    properties:
      accountId:
        type: string
    type: object
  v1.AccountRequestBody:
    properties:
      args:
        items:
          $ref: '#/definitions/v1.AccountIdBody'
        type: array
    type: object
...
info:
  contact:
    email: zoujh99@qq.com
    name: 夜央 Oh oh oh oh oh oh
  description: 基于区块链技术的房地产交易系统api文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 基于区块链技术的房地产交易系统api文档
  version: "1.0"
paths:
...
  /api/v1/hello:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 测试输出Hello
  /api/v1/queryAccountList:
    post:
      parameters:
      - description: account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/v1.AccountRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取账户信息
...
swagger: "2.0"